phase-8-advanced-compose/api/requirements.txt
flask==3.1.0
psycopg2-binary==2.9.9
redis==5.0.8
